+++ packages/gafl-webapp-service/src/pages/contact/contact/route.js
@@ -27,35 +27,45 @@ export const getData = async request => {
     throw new GetDataRedirect(LICENCE_LENGTH.uri)
   }
 
+  const junior = hasJunior(permission)
+
   return {
-    title: getTitle(permission, mssgs),
+    title: getTitle(permission, mssgs, junior),
     postHint: getPostHint(permission, mssgs),
-    content: getContent(permission, mssgs),
+    content: getContent(permission, mssgs, junior),
     emailConfirmation: permission.licensee.preferredMethodOfConfirmation === HOW_CONTACTED.email,
     emailText: getEmailText(permission, mssgs),
     mobileConfirmation: permission.licensee.preferredMethodOfConfirmation === HOW_CONTACTED.text,
     mobileText: getMobileText(permission, mssgs),
     licensee: permission.licensee,
     isPhysical: isPhysical(permission),
-    isJunior: hasJunior(permission)
+    errorMessage: getErrorText(permission, mssgs, junior)
   }
 }
 
-const getTitle = (permission, messages) =>
-  permission.isLicenceForYou
-    ? messages.important_info_contact_title_you
-    : messages.important_info_contact_title_other
+const getTitle = (permission, messages, junior) => {
+  if (permission.licenceLength === '12M' && !junior) {
+    return permission.isLicenceForYou ? messages.important_info_contact_title_you : messages.important_info_contact_title_other
+  }
+  return permission.isLicenceForYou ? messages.licence_confirm_method_where_title_you : messages.licence_confirm_method_where_title_other
+}
 
 const getPostHint = (permission, messages) =>
-  permission.isLicenceForYou
-    ? messages.important_info_contact_post_hint_you
-    : messages.important_info_contact_post_hint_other
+  permission.isLicenceForYou ? messages.important_info_contact_post_hint_you : messages.important_info_contact_post_hint_other
 
-const getContent = (permission, messages) => {
-  if (permission.licenceType === 'Salmon and sea trout') {
-    return permission.isLicenceForYou ? messages.important_info_contact_post_salmon_you : messages.important_info_contact_post_salmon_other
+const getContent = (permission, messages, junior) => {
+  const isSalmonLicense = permission.licenceType === 'Salmon and sea trout'
+  if (permission.licenceLength === '12M' && !junior) {
+    if (isSalmonLicense) {
+      return permission.isLicenceForYou
+        ? messages.important_info_contact_post_salmon_you
+        : messages.important_info_contact_post_salmon_other
+    }
+    return permission.isLicenceForYou
+      ? messages.important_info_contact_post_not_salmon_you
+      : messages.important_info_contact_post_not_salmon_other
   }
-  return permission.isLicenceForYou ? messages.important_info_contact_post_not_salmon_you : messages.important_info_contact_post_not_salmon_other
+  return isSalmonLicense ? messages.important_info_contact_content_salmon : messages.important_info_contact_content_not_salmon
 }
 
 const getMobileText = (permission, messages) =>
@@ -67,6 +77,11 @@ const getEmailText = (permission, messages) =>
   permission.licensee.preferredMethodOfConfirmation === HOW_CONTACTED.email
     ? `${messages.important_info_contact_item_email_value}${permission.licensee.email}`
     : messages.important_info_contact_item_email
+
+const getErrorText = (permission, messages, junior) =>
+  permission.licenceLength === '12M' && !junior
+    ? messages.important_info_contact_error_choose
+    : messages.important_info_contact_error_choose_short
 
 export const validator = Joi.object({
   'how-contacted': Joi.string().valid('email', 'text', 'none').required(),